+++ model_arena/model_arena/extensions/evaluation/evaluator.py
@@ -1,11 +1,17 @@
+import os
 import re
 import time
+import json
+import string
+import random
+import subprocess
 import pandas as pd
 
 from tqdm.auto import tqdm
 from abc import abstractmethod
 
 from ..base import BaseExtension
+from ...utils import extract_code_snippets
 
 from typing import Callable
 from pandas.core.series import Series
@@ -50,14 +56,20 @@ class BaseEvaluator(BaseExtension):
     def _evaluate(self, df: DataFrame) -> DataFrame:
         ...
 
-    def evaluate(self, df: DataFrame) -> DataFrame:
+    @abstractmethod
+    def _post_evaluate(self, df: DataFrame) -> tuple[DataFrame, DataFrame]:
+        ...
+
+    def evaluate(self, df: DataFrame) -> tuple[DataFrame, DataFrame]:
         # preprocess dataframe with _ignore and _transform
         df = self._preprocess(df)
         # evaluate
         df.loc[:, "auditor"] = self.extension_name
         df = self._evaluate(df)
+        # post evaluate, split success and failed
+        success_df, failed_df = self._post_evaluate(df)
 
-        return df
+        return success_df, failed_df
 
 
 class UnaryEvaluator(BaseEvaluator):
@@ -80,6 +92,70 @@ class ExactMatchEvaluator(UnaryEvaluator):
     def _evaluate(self, df: DataFrame) -> DataFrame:
         df["score"] = df.progress_apply(lambda x: int(x["output"] == x["label"]), axis=1)
         return df
+
+    def _post_evaluate(self, df: DataFrame) -> tuple[DataFrame, DataFrame]:
+        return df, pd.DataFrame(columns=df.columns)
+
+
+class PythonExecutionEvaluator(UnaryEvaluator):
+    extension_name: str = "python_execution"
+
+    def _is_passed(self, instruction: str, output: str, test: str) -> bool:
+        code = f"{instruction}{output}\n{test}"
+        test_file = f"{''.join(random.choices(string.ascii_lowercase, k=10))}.py"
+
+        with open(test_file, "w", encoding="utf-8") as f:
+            f.write(code)
+
+        try:
+            subprocess.run(["python", f"{test_file}"], check=True, capture_output=True, text=True, timeout=5)
+            result = True
+        except Exception:
+            result = False
+
+        os.remove(test_file)
+
+        return result
+
+    def _evaluate_once(self, row: Series) -> Series:
+        instruction, output = row["instruction"], row["output"]
+        test = json.loads(row["information"])["test"]
+
+        # try to obtain the code solution if possible
+        # we only find the first code snippet wrapped in
+        # the markdown format
+        code_output = extract_code_snippets(output)
+        if len(code_output) > 0:
+            try:
+                # this should find the proper code solution
+                # if not we may only extract the language successfully
+                # not the code itself
+                code_output = code_output[0][1]
+            except Exception:
+                # fall back to blank code output
+                code_output = ""
+        else:
+            # we don not find any code output
+            code_output = ""
+
+        # HACK: this is a hack to deal with the different type of model
+        # we do not know if the model is a continual writing model or a chat model
+        # so we check both possible solution
+        passed = (
+            self._is_passed(instruction, output, test)
+            or self._is_passed("", output, test)
+            or self._is_passed(instruction, code_output, test)
+            or self._is_passed("", code_output, test)
+        )
+
+        return pd.Series([int(passed)], index=["score"])
+
+    def _evaluate(self, df: DataFrame) -> DataFrame:
+        df["score"] = df.progress_apply(self._evaluate_once, axis=1)
+        return df
+
+    def _post_evaluate(self, df: DataFrame) -> tuple[DataFrame, DataFrame]:
+        return df, pd.DataFrame(columns=df.columns)
 
 
 class ChatGPTEvaluator(PairwiseEvaluator):
@@ -104,13 +180,15 @@ class ChatGPTEvaluator(PairwiseEvaluator):
             prompt=prompt if prompt else DEFAULT_PROMPT,
             num_retries=num_retries,
         )
-        self.llm = BytedChatGPT(model=self.model)
+        self.llm = BytedChatGPT(model_name=self.model)
         self.extension_name = self.extension_name.format(model=self.model)
+        # we set a wait time to avoid high qps request
+        self._time_elapsed = 2
 
     def _evaluate_once(self, row: Series) -> Series:
         for _ in range(self.num_retries):
-            # avoid high qps for gpt requests
-            time.sleep(2)
+            # time waited to send another request
+            time.sleep(self._time_elapsed)
             try:
                 results = self.llm.invoke(
                     self.prompt.format(
@@ -119,6 +197,9 @@ class ChatGPTEvaluator(PairwiseEvaluator):
                         output_y=row["output_y"],
                     )
                 )
+                # we successfully send a gpt request
+                # we reset the wait time
+                self._time_elapsed = 2
                 if scores := re.search(r"\(\d,[ ]?\d\)", results)[0]:
                     return pd.Series(
                         tuple(map(int, scores[1:-1].replace(", ", ",").split(","))),
@@ -127,6 +208,8 @@ class ChatGPTEvaluator(PairwiseEvaluator):
                 else:
                     continue
             except Exception:
+                # we failed to send a gpt request, we make wait time longer
+                self._time_elapsed += 5
                 continue
 
         return pd.Series((-1, -1), index=["score_x", "score_y"])
@@ -134,3 +217,10 @@ class ChatGPTEvaluator(PairwiseEvaluator):
     def _evaluate(self, df: DataFrame) -> DataFrame:
         df[["score_x", "score_y"]] = df.progress_apply(self._evaluate_once, axis=1)
         return df
+
+    def _post_evaluate(self, df: DataFrame) -> tuple[DataFrame, DataFrame]:
+        failed_pattern = (df["score_x"] == -1) & (df["score_y"] == -1)
+        success_df = df[~failed_pattern]
+        failed_df = df[failed_pattern]
+
+        return success_df, failed_df