+++ model_arena/model_arena/utils.py
@@ -1,0 +1,6 @@
+import re
+
+
+def extract_code_snippets(s: str) -> list[tuple[str, str]]:
+    code_snippets = re.findall(r"```(\w+?)\n(.*?)```", s, re.DOTALL)
+    return code_snippets